requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
langchain_chroma
beautifulsoup4
faiss-cpu
streamlit
langchain_groq
langchain_core